requirements.txt
molecule==4.0.4
ansible==8.7.0
gitpython==3.1.43
# https://pip.pypa.io/en/stable/topics/vcs-support/